TRABAIO FINALIZEIDO
TERMINAMOS TUDO, TA PERFEITO

+++ Trabalho/Trabalho_Compiladores_Artur_Eduardo_Pedro/src/main/java/com/mycompany/trabalho_analisador_sintatico_artur_pedro_ryudi/Analisador_semantico.java
@@ -111,7 +111,7 @@ public class Analisador_semantico implements Analisador_sintaticoConstants {
     return null; // erro
   }
 
-  public void addtabelaSimbolos(String nome, String categoria,float atributo){
+  public void addtabelaSimbolos(String nome, String categoria,int atributo){
     TabelaSimbolos.add(new Simbolo(nome,categoria,atributo));
   }
